Added function to find p_star for Riemann problem
Added calculate_pstar, which solves for p_star (the pressure on either
side of the contact surface in the general Riemann problem). This solves
the equations shown in Knight using Newton's Method. One issue is that
the initial guess currently is not very smart, better ones exist in the
literature. Another issue is that the Newton's Method solver runs for a
fixed number of iterations instead of stopping once an error tolerance
is reached.