First steps in complete restructuring of code
Added new classes, Flux and Algebra, that will be used in calculating
fluxes and doing linear algebra operations. This will make the code more
versatile and able to implement other methods. Began work on
implementing Steger and Warming's method for calculating wall fluxes.
Deleted large amounts of previous code. Note that this commit does not
compile and is the first in a series of commits to fix this code.