Organized project directory and added Makefile
The project is now split into directories for the source code, headers,
objects, and executables. A Makefile in the root directory now compiles
the source from the src directory into the bin directory. The code now
successfully compiles with Make but contains many bugs still and does
not yet work properly.