Fixed eigenvectors, mostly working code
Fixed left eigenvectors to be correct (they are now the inverse of the
right eigenvectors). Results from the code are still wrong though, the
shock eventually stops in the tube and there is no expansion behind it.

The Makefile is now set to compile with -O3, since without that flag the
results from running the code are different every time, even with no
changes whatsoever and without recompiling. This makes no sense anyways
since the code is not parallelized, but -O3 seems to fix this. The root
cause of this bug is still unknown.