Bug fixes and minor improvements
Fix a bug where the q vector was rewriting itself every iteration, which
caused overlaps between rewritten and non-rewritten parts of the
flowfield. The code now copies the q vector at the beginning of an
iteration into a q_old vector and uses that to generate q.

Fixed issues with grid spacing/initialization where it ignored ghost
cells. This also required some changes to the indices in animation.py,
since it now needs to read the ghost cells.

Added a simple build script, build.sh, which cleans the directory,
builds the code and the tests, then runs the tests.