More refactoring, initialized cells and solution
Cells are now initialized in an unordered_map within flowfield.cpp.
Removed constructor of Cell/Ghost, since forward declaring an
unordered_map<int, Cell> in flowfield.h causes the default constructor
of Cell to be called, Cell::Cell(). This is an issue since Cell has no
default constructor, and we need the members of Cell objects to be
initialized so a default constructor would not be useful. Two solutions
exist, one would be to create an unordered_map<int, *Cell> to avoid the
constructor and the other was to get rid of the constructor and
initialize members of Cell within Flowfield (since that is where those
values are calculated anyways). Went with the latter option since I
believe it is simpler and cleaner.