Added Flux and Face class, more refactoring
This commit is significant. Added the Flux class, which contains a
static method for computing the flux across a face with Steger and
Warming's method. Also added the Face class and initialized some faces
between cells in flowfield.cpp. Also added a python script to attempt to
derive the equations for the fluxes for Steger and Warming's method, but
it is incomplete.