Changed some class names, started using sets
Changed name of abstract base class Cell to Volume, and changed name of
child class Flow to Cell. Placed cells, ghosts, volumes, and faces into
unordered_set containers alongside the unordered_maps, so that iteration
can be done without referencing the key, only the value. This update
made the code significantly slower though, so the changes may not be
permanent. Code must be profiled in comparison to older version to check
for where speed loss is happening.