Added HDF5 for reading in CGNS meshes
Added and linked the HDF5 library and wrote some basic code in
inputs.cpp for reading a single piece of data from a .cgns file. It
works, so the next step is to get the code to read the whole mesh and
boundary conditions and connectivity in.

Also, there needs to be an automated way to get HDF5 for other users,
since it is too large to include in the repository. Consider adding a
script that builds HDF5 automatically.