Many changes, boundary conditions, time integrator
Face fluxes are now (probably) correctly oriented so that normal vectors
always point towards the cell with higher cell ID. This has not been
tested. Made progress on overhauling ghost cells and boundary
conditions, and added the BC enum for boundary condition types.
Currently, inflow and wall are attempting to be implemented, but they
are definitely not done.