Code now passes the constant box test
Initializing in a domain with constant density and pressure with zero
velocity now gives an unchanging result, where density and pressure stay
the same and velocity stays at zero. This is also true when a mixture of
inflow, outflow, and wall BCs are used (if the inflows are also the same
properties as the volume) so this is good. So far, this has been
top-level tested with the box, rhombus, and wedge_test meshes. This will
need to be organized into a high-level unit test soon.