Progress on writing results in CGNS format
Trying to plot results with Python was getting clunky with 2D
unstructured quadrilaterals (Python prefers either structured, or
unstructured triangles) so the efforts are shifting to writing results
out in CGNS format. The goal is to get the results written to the same
file as the original mesh, to have a sort of combined case file where
everything is stored. This is being implemented in the Output class.