First (probably) working commit of Shock, more IO
This is the first commit where wedge_30deg.cgns can be read in and ran,
with the solution being read into Tecplot. There are many caveats to
this, for example the data output is only ASCII right now, the solution
was not run until convergence, and solution was not compared to oblique
shock solutions, and it took very long to run. However, this will be a
good starting point. The main issue now is the mesh generation takes
very long since every face requires iterating over the whole set of
faces to check for existence, and this really sucks. Will be working on
this problem.