Finished face initialization algorithm (for now)
The faces are now constructed for every cell, with duplicates, and the
duplicates are combined as a second step afterwords. It seems to work
and is much faster than before, however the code looks pretty rough
right now and needs to be refactored and unit tested. Also, flow over a
wedge is giving strange results (maybe the wall BC is wrong?)