Added TimeIntegrator class, Forward Euler
The time integrator is now more modular, since the Forward Euler method
is now a function part of the TimeIntegrator static class. This way,
other time integrators can easily be added. Also, modified the Makefile
to slightly change the HDF5 path. Also, got rid of the update_ghosts
function in Flowfield and included the contents as part of
apply_time_integrator.