Refactored TimeIntegrator, added ForwardEuler
The TimeIntegrator class is now an abstract base class that is inherited
by other specific time integrator classes, the only implementation at
the moment being ForwardEuler. The TimeIntegrator class now has a static
function choose_time_integrator that is called in main that generates a
TimeIntegrator* pointer, but the function instantiates an object of the
correct class depending on the input choice of time integrator. This
lets us choose the time integrator with run-time polymorphism.